requirements.txt
vosk
ffmpeg-python
watchdog
pyzmq
